Crate dlwp

dlwp

DarkLight Web Protocol Library

This library is used for interacting with darklight_driver to create or connect to DarkLight streams and send or receive Messages. Documentation can be found on docs.rs, examples can be found in the tests and test_streams/

Requirements

As mentioned in the darklight_driver setup, libudev is needed to compile on Linux OSes if the use_io feature is enabled.

Features in Cargo.toml:

[features]
# Enable if testing a stream
test_stream = []
# If enabled the serde and serde_json crates will be publicly available from dlwp
include_serde = ["serde/serde_derive", "serde/std"]
# If enabled the chrono crate will be publicly available from dlwp
include_chrono = ["chrono"]
# Used by DarkLight driver, this is not needed for regular use
use_io = ["serialport"]

Re-exports

pub extern crate cerpton;

Modules

codes

“codes” for specifying information in a Message

config

Configurations for DarkLight

connections

For Connections struct and handling of client connections (for servers)

encryption

Type for encryption

id

Contains types and functions for using different types of Ids

langs

message

For a Message

stream

Used for creating DarkLight clients and servers