Module dlwp::codes

“codes” for specifying information in a Message

Enums

Code

Constants

CHUNK_RESPONSE

Responed with a chunk of a large piece of data, should coem after a chunk has been requested

CONNECTION_ACCEPTED

Response for a connection being accepted, should be sent after requesting a connection

CONNECTION_DENIED

Response for when a connection is being denied

DISCONNECT

Disconnect server/client

DISTRIBUTOR_NOT_FOUND

Internal error indicating that the distributor could not be found

FILE_RESPONSE

Response for a file that fits inside a Message or wether the file can be sent or not. If the file needs to be sent in chunks respond with the number of chunks that will need to be received

INVALID_DID

Internal error indicating that the Distributor Id is not valid

INVALID_RR

When a client requests or responds with something unexpected

LENGTH_EXCEEDED

Internal error indicating that a Message will be longer than 4096 bytes

READ_FAILED

READ_SUCCESS

READ_TIMEDOUT

REGULAR_RESPONSE

Response for any type of request, should be sent after requesting a “regular” response

REMOVE_CLIENT

REQUEST_CHUNK

Request a chunk of a large piece of data, usually comes after specifying the type of data

REQUEST_CONNECTION

Requests a connection from the Server or Client

REQUEST_FILE

Requests a file, if the file can fit into a Message it should be sent in the next response, if not it should be sent in chunks

REQUEST_PORT_SWITCH

Requests if changing Server/Client ports will be acceptable

REQUEST_RESPONSE

Requests that the server or client responds

STATUS_OK

TEAPOT

Do what you want with this

UNKNOWN_STATUS

WRITE_FAILED

WRITE_SUCCESS

WRITE_TIMEDOUT