Module dlwp::message

For a Message

Structs

Message

ReceiveInfo

Struct containing information about the receiver

TransmitInfo

Struct containing information about the transmitter

Constants

CONTENTS_LENGTH

INFO_SPLIT

Divides the Message when it is represented as a String

MSG_END

The end of a message being sent

MSG_INIT

The start of a message being sent

SN_MSG_INIT

The start of a message being sent on a serial device

Functions

contents_to_string

If the original String was less than 4096 bytes there will be many null characters (\0) at the end

split_from_info

Splits a String using INFO_SPLIT into an &str array. Has inline incase it is called regularaly.

string_to_contents

Converts a String to Contents, will panic if the input String is larger than 4096 bytes

valid_message_string

Checks if a Message represented as a String is valid. This starts by checking if it’s “info splits” are in the right places. If it’s ReceiveInfo and TransmitInfo does not contain empty values.

Type Aliases

Contents

“contents” type, 4096 byte array